pyproject.toml
[project]
name = "cli-sandbox"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12, <3.14"
dependencies = [
    "typer[all] (>=0.15.3,<0.16.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]

[tool.poetry]
packages = [{include = "cli_sandbox", from = "src"}]

[tool.poetry.group.dev.dependencies]
pyinstaller = "^6.13.0"
pytest = "^8.3.5"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
cli = "cli_sandbox.main:app"